+++ src/bittytax/conv/parsers/bscscan.py
@@ -108,10 +108,10 @@ bsc_txns = DataParser(
     ParserType.EXPLORER,
     "BscScan (BSC Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "ContractAddress",
@@ -123,7 +123,7 @@ bsc_txns = DataParser(
         "Historical $Price/BNB",
         "Status",
         "ErrCode",
-        "Method",
+        "Method",  # New field
     ],
     worksheet_name=WORKSHEET_NAME,
     row_handler=parse_bscscan,
@@ -133,10 +133,10 @@ DataParser(
     ParserType.EXPLORER,
     "BscScan (BSC Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "ContractAddress",
@@ -148,7 +148,7 @@ DataParser(
         "Historical $Price/BNB",
         "Status",
         "ErrCode",
-        "Method",
+        "Method",  # New field
         "PrivateNote",
     ],
     worksheet_name=WORKSHEET_NAME,
@@ -208,10 +208,10 @@ bsc_int = DataParser(
     ParserType.EXPLORER,
     "BscScan (BSC Internal Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "ParentTxFrom",
         "ParentTxTo",
         lambda c: c in ("ParentTxETH_Value", "ParentTxBNB_Value"),
@@ -234,10 +234,10 @@ DataParser(
     ParserType.EXPLORER,
     "BscScan (BSC Internal Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "ParentTxFrom",
         "ParentTxTo",
         lambda c: c in ("ParentTxETH_Value", "ParentTxBNB_Value"),