+++ src/bittytax/conv/parsers/etherscan.py
@@ -193,10 +193,10 @@ etherscan_txns = DataParser(
     ParserType.EXPLORER,
     "Etherscan (ETH Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "ContractAddress",
@@ -208,7 +208,7 @@ etherscan_txns = DataParser(
         "Historical $Price/Eth",
         "Status",
         "ErrCode",
-        "Method",
+        "Method",  # New field
     ],
     worksheet_name="Etherscan",
     row_handler=parse_etherscan,
@@ -218,10 +218,10 @@ DataParser(
     ParserType.EXPLORER,
     "Etherscan (ETH Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "ContractAddress",
@@ -233,7 +233,7 @@ DataParser(
         "Historical $Price/Eth",
         "Status",
         "ErrCode",
-        "Method",
+        "Method",  # New field
         "PrivateNote",
     ],
     worksheet_name="Etherscan",
@@ -293,10 +293,10 @@ etherscan_int = DataParser(
     ParserType.EXPLORER,
     "Etherscan (ETH Internal Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "ParentTxFrom",
         "ParentTxTo",
         "ParentTxETH_Value",
@@ -319,10 +319,10 @@ DataParser(
     ParserType.EXPLORER,
     "Etherscan (ETH Internal Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "ParentTxFrom",
         "ParentTxTo",
         "ParentTxETH_Value",
@@ -346,10 +346,10 @@ etherscan_tokens = DataParser(
     ParserType.EXPLORER,
     "Etherscan (Token Transfers ERC-20)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",  # New field
         "UnixTimestamp",
-        lambda c: c in ("DateTime", "DateTime (UTC)"),
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "TokenValue",  # Renamed
@@ -384,7 +384,7 @@ etherscan_nfts = DataParser(
     ParserType.EXPLORER,
     "Etherscan (NFT Transfers ERC-721 & ERC-1155)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
         "DateTime (UTC)",  # Renamed