+++ src/bittytax/conv/parsers/hecoinfo.py
@@ -108,10 +108,10 @@ heco_txns = DataParser(
     ParserType.EXPLORER,
     "HecoInfo (HECO Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        "DateTime",
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "ContractAddress",
@@ -123,6 +123,7 @@ heco_txns = DataParser(
         "Historical $Price/HT",
         "Status",
         "ErrCode",
+        "Method",  # New field
     ],
     worksheet_name=WORKSHEET_NAME,
     row_handler=parse_hecoinfo,
@@ -132,10 +133,10 @@ DataParser(
     ParserType.EXPLORER,
     "HecoInfo (HECO Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        "DateTime",
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "From",
         "To",
         "ContractAddress",
@@ -147,6 +148,7 @@ DataParser(
         "Historical $Price/HT",
         "Status",
         "ErrCode",
+        "Method",  # New field
         "PrivateNote",
     ],
     worksheet_name=WORKSHEET_NAME,
@@ -172,7 +174,6 @@ DataParser(
         "Historical $Price/HT",
         "Status",
         "ErrCode",
-        "Method",
     ],
     worksheet_name=WORKSHEET_NAME,
     row_handler=parse_hecoinfo,
@@ -197,7 +198,6 @@ DataParser(
         "Historical $Price/HT",
         "Status",
         "ErrCode",
-        "Method",
         "PrivateNote",
     ],
     worksheet_name=WORKSHEET_NAME,
@@ -208,10 +208,10 @@ heco_int = DataParser(
     ParserType.EXPLORER,
     "HecoInfo (HECO Internal Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        "DateTime",
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "ParentTxFrom",
         "ParentTxTo",
         "ParentTxETH_Value",
@@ -234,10 +234,10 @@ DataParser(
     ParserType.EXPLORER,
     "HecoInfo (HECO Internal Transactions)",
     [
-        "Txhash",
+        lambda c: c in ("Txhash", "Transaction Hash"),  # Renamed
         "Blockno",
         "UnixTimestamp",
-        "DateTime",
+        lambda c: c in ("DateTime", "DateTime (UTC)"),  # Renamed
         "ParentTxFrom",
         "ParentTxTo",
         "ParentTxETH_Value",